Sub: Remove unsused aux switch code
