+++ ArduSub/AP_State.cpp
@@ -74,12 +74,6 @@ void Sub::set_pre_arm_check(bool b)
     }
 }
 
-void Sub::update_using_interlock()
-{
-    // check if we are using motor interlock control on an aux switch
-    ap.using_interlock = check_if_auxsw_mode_used(AUXSW_MOTOR_INTERLOCK);
-}
-
 void Sub::set_motor_emergency_stop(bool b)
 {
     if (ap.motor_emergency_stop != b) {
