+++ ArduSub/config.h
@@ -80,14 +80,6 @@
 #endif
 
 //////////////////////////////////////////////////////////////////////////////
-// Aux Switches
-//
-
-#ifndef AUXSW_ENABLED
-# define AUXSW_ENABLED DISABLED
-#endif
-
-//////////////////////////////////////////////////////////////////////////////
 // RPM
 //
