+++ ArduSub/defines.h
@@ -29,61 +29,6 @@ enum autopilot_yaw_mode {
     AUTO_YAW_CORRECT_XTRACK =    6   // steer the sub in order to correct for crosstrack error during line following
 };
 
-// Ch6... Ch12 aux switch control
-#define AUX_SWITCH_PWM_TRIGGER_HIGH 1800   // pwm value above which the ch7 or ch8 option will be invoked
-#define AUX_SWITCH_PWM_TRIGGER_LOW  1200   // pwm value below which the ch7 or ch8 option will be disabled
-#define CH6_PWM_TRIGGER_HIGH    1800
-#define CH6_PWM_TRIGGER_LOW     1200
-
-// values used by the ap.ch7_opt and ap.ch8_opt flags
-#define AUX_SWITCH_LOW              0       // indicates auxiliary switch is in the low position (pwm <1200)
-#define AUX_SWITCH_MIDDLE           1       // indicates auxiliary switch is in the middle position (pwm >1200, <1800)
-#define AUX_SWITCH_HIGH             2       // indicates auxiliary switch is in the high position (pwm >1800)
-
-// Aux Switch enumeration
-enum aux_sw_func {
-    AUXSW_DO_NOTHING =           0, // aux switch disabled
-    //    AUXSW_FLIP =                 2, // flip
-    AUXSW_SIMPLE_MODE =          3, // change to simple mode
-
-    // No RTL mode for Sub
-    //    AUXSW_RTL =                  4, // change to RTL flight mode
-
-    AUXSW_SAVE_TRIM =            5, // save current position as level
-    AUXSW_SAVE_WP =              7, // save mission waypoint or RTL if in auto mode
-    AUXSW_CAMERA_TRIGGER =       9, // trigger camera servo or relay
-    AUXSW_RANGEFINDER =         10, // allow enabling or disabling rangefinder in flight which helps avoid surface tracking when you are far above the ground
-    AUXSW_FENCE =               11, // allow enabling or disabling fence in flight
-    AUXSW_RESETTOARMEDYAW =     12, // changes yaw to be same as when quad was armed
-    AUXSW_SUPERSIMPLE_MODE =    13, // change to simple mode in middle, super simple at top
-    AUXSW_ACRO_TRAINER =        14, // low = disabled, middle = leveled, high = leveled and limited
-
-    // No sprayer for Sub, remove
-    //    AUXSW_SPRAYER =             15, // enable/disable the crop sprayer
-
-    AUXSW_AUTO =                16, // change to auto flight mode
-    AUXSW_GRIPPER =             19, // Operate cargo grippers low=off, middle=neutral, high=on
-
-    // No parachute for Sub, remove
-    //    AUXSW_PARACHUTE_ENABLE  =   21, // Parachute enable/disable
-    //    AUXSW_PARACHUTE_RELEASE =   22, // Parachute release
-    //    AUXSW_PARACHUTE_3POS =      23, // Parachute disable, enable, release with 3 position switch
-
-    AUXSW_MISSION_RESET =       24, // Reset auto mission to start from first command
-    AUXSW_ATTCON_FEEDFWD =      25, // enable/disable the roll and pitch rate feed forward
-    AUXSW_ATTCON_ACCEL_LIM =    26, // enable/disable the roll, pitch and yaw accel limiting
-    AUXSW_RETRACT_MOUNT =       27, // Retract Mount
-    AUXSW_RELAY =               28, // Relay pin on/off (only supports first relay)
-
-    AUXSW_LOST_VEHICLE_SOUND =   30, // Play lost vehicle sound
-    AUXSW_MOTOR_ESTOP =         31, // Emergency Stop Switch
-    AUXSW_MOTOR_INTERLOCK =     32, // Motor On/Off switch
-    //    AUXSW_BRAKE =               33, // Brake flight mode
-    AUXSW_RELAY2 =              34, // Relay2 pin on/off (in Mission planner set CH8_OPT  = 34)
-    AUXSW_RELAY3 =              35, // Relay3 pin on/off (in Mission planner set CH9_OPT  = 35)
-    AUXSW_RELAY4 =              36, // Relay4 pin on/off (in Mission planner set CH10_OPT = 36)
-};
-
 // HIL enumerations
 #define HIL_MODE_DISABLED               0
 #define HIL_MODE_SENSORS                1
